+++ src/spell/package.json
@@ -11,7 +11,6 @@
     "debug": "~2.6.9",
     "express": "~4.16.1",
     "fast-json-stringify": "^1.16.2",
-    "hanspell": "^0.7.4",
     "html-entities": "^1.2.1",
     "http-errors": "~1.6.3",
     "jade": "~1.11.0",
